testing blob deletion

+++ netlify/edge-functions/blobs.ts
@@ -15,14 +15,16 @@ export default async (req: Request, context: Context) => {
     });
 
   } else {
-    const id = Date.now().toString();
-    const prefix = id.slice(0, 6);
-    const { blobs } = await collection.list({ prefix });
+    const { blobs } = await collection.list();
 
     blobs.forEach(blob => {
-      collection.delete(blob.key)
+      const oldDate = parseInt(blob.key);
+      if (isOneWeekOld(oldDate))
+        collection.delete(blob.key)
     });
+
     const data = await req.json();
+    const id = Date.now().toString();
     const link = context.url.origin + '/list?blob=' + id;
 
     await collection.setJSON(id, data);
@@ -38,3 +40,11 @@ export default async (req: Request, context: Context) => {
 export const config: Config = {
   path: ['/blob', '/blob/:uid'],
 };
+
+function isOneWeekOld(oldDate: number) {
+  const oneWeekInMilliseconds = 60 * 60 * 1000;
+  const currentDate = Date.now();
+  const timeDifference = currentDate - oldDate;
+
+  return timeDifference <= oneWeekInMilliseconds;
+}
